requirements.txt
nextcord
psutil 
python-dotenv 
art 
cowsay
rich
yt-dlp
aiohttp[speedups]